+++ library/core/src/main/java/com/google/android/exoplayer2/source/MaskingMediaSource.java
@@ -22,7 +22,6 @@ import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.MediaItem;
 import com.google.android.exoplayer2.Timeline;
 import com.google.android.exoplayer2.Timeline.Window;
-import com.google.android.exoplayer2.source.MediaSourceEventListener.EventDispatcher;
 import com.google.android.exoplayer2.upstream.Allocator;
 import com.google.android.exoplayer2.upstream.TransferListener;
 import com.google.android.exoplayer2.util.Assertions;
@@ -42,7 +41,6 @@ public final class MaskingMediaSource extends CompositeMediaSource<Void> {
 
   private MaskingTimeline timeline;
   @Nullable private MaskingMediaPeriod unpreparedMaskingMediaPeriod;
-  @Nullable private EventDispatcher unpreparedMaskingMediaPeriodEventDispatcher;
   private boolean hasStartedPreparing;
   private boolean isPrepared;
   private boolean hasRealTimeline;
@@ -121,9 +119,6 @@ public final class MaskingMediaSource extends CompositeMediaSource<Void> {
       // unset and we don't load beyond periods with unset duration. We need to figure out how to
       // handle the prepare positions of multiple deferred media periods, should that ever change.
       unpreparedMaskingMediaPeriod = mediaPeriod;
-      unpreparedMaskingMediaPeriodEventDispatcher =
-          createEventDispatcher(/* windowIndex= */ 0, id, /* mediaTimeOffsetMs= */ 0);
-      unpreparedMaskingMediaPeriodEventDispatcher.mediaPeriodCreated();
       if (!hasStartedPreparing) {
         hasStartedPreparing = true;
         prepareChildSource(/* id= */ null, mediaSource);
@@ -136,8 +131,6 @@ public final class MaskingMediaSource extends CompositeMediaSource<Void> {
   public void releasePeriod(MediaPeriod mediaPeriod) {
     ((MaskingMediaPeriod) mediaPeriod).releasePeriod();
     if (mediaPeriod == unpreparedMaskingMediaPeriod) {
-      Assertions.checkNotNull(unpreparedMaskingMediaPeriodEventDispatcher).mediaPeriodReleased();
-      unpreparedMaskingMediaPeriodEventDispatcher = null;
       unpreparedMaskingMediaPeriod = null;
     }
   }
@@ -220,14 +213,6 @@ public final class MaskingMediaSource extends CompositeMediaSource<Void> {
   protected MediaPeriodId getMediaPeriodIdForChildMediaPeriodId(
       Void id, MediaPeriodId mediaPeriodId) {
     return mediaPeriodId.copyWithPeriodUid(getExternalPeriodUid(mediaPeriodId.periodUid));
-  }
-
-  @Override
-  protected boolean shouldDispatchCreateOrReleaseEvent(MediaPeriodId mediaPeriodId) {
-    // Suppress create and release events for the period created while the source was still
-    // unprepared, as we send these events from this class.
-    return unpreparedMaskingMediaPeriod == null
-        || !mediaPeriodId.equals(unpreparedMaskingMediaPeriod.id);
   }
 
   private Object getInternalPeriodUid(Object externalPeriodUid) {